Implement binary marshal/unmarshalers
In order to serialise a `Source` properly too and from redis, we must
implement these interfaces. We could, instead, keep sources as naked
`uint8`s and, instead, manually assign numeric values but that's messier
and takes away the power later to be able to put other operations on
Sources, like strings or workflow rules.